[refactor] 모듈 분리 (파일 관리하는 모듈 생성)
